Refactor haskell.nix expressions
The single source file was moved to the top-level of the repository.

The haskell.nix expressions have been updated to use a pin common with
all my projects. Reasoning being most of the deps will be compiled with
the newer pin anyway so all projects might as well use that.